improving docs
