+++ docs/src/pages/AccordionDemo.tsx
@@ -36,9 +36,7 @@ export default function AccordionDemoPage() {
   return (
     <Surface>
       <NavDrawer />
-      <Stack
-        preset="showcaseStack"
-      >
+      <Stack>
         {/* Header --------------------------------------------------------- */}
         <Typography variant="h2" bold>
           Accordion Showcase
