+++ docs/src/pages/DateSelectorDemo.tsx
@@ -87,7 +87,7 @@ export default function DateSelectorDemoPage() {
   return (
     <Surface>
       <NavDrawer />
-      <Stack preset="showcaseStack">
+      <Stack>
         <Typography variant="h2" bold>
           DateSelector Showcase
         </Typography>
