+++ docs/src/pages/IconButtonDemoPage.tsx
@@ -104,9 +104,7 @@ export default function IconButtonDemoPage() {
   return (
     <Surface>
       <NavDrawer />
-      <Stack
-        preset="showcaseStack"
-      >
+      <Stack>
         {/* Page header ----------------------------------------------------- */}
         <Typography variant="h2" bold>
           IconButton Showcase
