+++ docs/src/pages/RadioGroupDemo.tsx
@@ -131,9 +131,7 @@ export default function RadioGroupDemoPage() {
   return (
     <Surface>
       <NavDrawer />
-      <Stack
-        preset="showcaseStack"
-      >
+      <Stack>
         {/* Page header ----------------------------------------------------- */}
         <Typography variant="h2" bold>
           Radio Group Showcase
