+++ docs/src/pages/SnackbarDemo.tsx
@@ -42,7 +42,7 @@ export default function SnackbarDemoPage() {
   return (
     <Surface>
       <NavDrawer />
-      <Stack preset="showcaseStack">
+      <Stack>
         {/* Header --------------------------------------------------------- */}
         <Typography variant="h2" bold>
           Snackbar Showcase
