+++ docs/tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/vite-env.d.ts","./src/presets/globalPresets.ts","./src/App.tsx","./src/main.tsx","./src/pages/AccordionConstrainedDemo.tsx","./src/pages/AccordionDemo.tsx","./src/pages/AppBarDemo.tsx","./src/pages/AvatarDemo.tsx","./src/pages/BoxDemo.tsx","./src/pages/ButtonDemoPage.tsx","./src/pages/ChatDemo.tsx","./src/pages/CheckBoxDemo.tsx","./src/pages/DrawerDemo.tsx","./src/pages/FormDemoPage.tsx","./src/pages/GridDemo.tsx","./src/pages/IconButtonDemoPage.tsx","./src/pages/IconDemoPage.tsx","./src/pages/Installation.tsx","./src/pages/ListDemoPage.tsx","./src/pages/MainPage.tsx","./src/pages/ModalDemo.tsx","./src/pages/Overview.tsx","./src/pages/PaginationDemo.tsx","./src/pages/PanelDemo.tsx","./src/pages/ParallaxDemo.tsx","./src/pages/PresetDemoPage.tsx","./src/pages/ProgressDemo.tsx","./src/pages/RadioGroupDemo.tsx","./src/pages/SelectDemo.tsx","./src/pages/SliderDemo.tsx","./src/pages/SnackbarDemo.tsx","./src/pages/SpeedDialDemo.tsx","./src/pages/StepperDemo.tsx","./src/pages/SwitchDemo.tsx","./src/pages/TableDemo.tsx","./src/pages/TabsDemo.tsx","./src/pages/Test.tsx","./src/pages/TextFormDemo.tsx","./src/pages/TooltipDemo.tsx","./src/pages/TreeDemo.tsx","./src/pages/TypographyDemoPage.tsx","./src/pages/Usage.tsx","./src/pages/VideoDemo.tsx"],"version":"5.8.3"}
+{"root":["./src/vite-env.d.ts","./src/presets/globalPresets.ts","./src/App.tsx","./src/main.tsx","./src/components/NavDrawer.tsx","./src/pages/AccordionConstrainedDemo.tsx","./src/pages/AccordionDemo.tsx","./src/pages/AppBarDemo.tsx","./src/pages/AvatarDemo.tsx","./src/pages/BoxDemo.tsx","./src/pages/ButtonDemoPage.tsx","./src/pages/CheckBoxDemo.tsx","./src/pages/DateSelectorDemo.tsx","./src/pages/DrawerDemo.tsx","./src/pages/FormDemoPage.tsx","./src/pages/GridDemo.tsx","./src/pages/IconButtonDemoPage.tsx","./src/pages/IconDemoPage.tsx","./src/pages/ImageDemo.tsx","./src/pages/Installation.tsx","./src/pages/IteratorDemo.tsx","./src/pages/ListDemoPage.tsx","./src/pages/MainPage.tsx","./src/pages/ModalDemo.tsx","./src/pages/OAIChatDemo.tsx","./src/pages/Overview.tsx","./src/pages/PaginationDemo.tsx","./src/pages/ParallaxDemo.tsx","./src/pages/PresetDemoPage.tsx","./src/pages/ProgressDemo.tsx","./src/pages/PropPatterns.tsx","./src/pages/RadioGroupDemo.tsx","./src/pages/SelectDemo.tsx","./src/pages/SliderDemo.tsx","./src/pages/SnackbarDemo.tsx","./src/pages/SpeedDialDemo.tsx","./src/pages/StackDemo.tsx","./src/pages/StepperDemo.tsx","./src/pages/SurfaceExplainer.tsx","./src/pages/SwitchDemo.tsx","./src/pages/TableDemo.tsx","./src/pages/TabsDemo.tsx","./src/pages/Test.tsx","./src/pages/TextFormDemo.tsx","./src/pages/TooltipDemo.tsx","./src/pages/TreeDemo.tsx","./src/pages/TypographyDemoPage.tsx","./src/pages/Usage.tsx","./src/pages/VideoDemo.tsx","./src/pages/components/Panel.tsx"],"version":"5.8.3"}
